streamlit_app/requirements.txt
streamlit==1.29.0
requests==2.31.0
pandas==2.1.4
plotly==5.17.0
python-dotenv==1.0.0